+++ modules/marshaller-common/src/main/java/org/apache/ignite/internal/marshaller/Marshaller.java
@@ -48,7 +48,16 @@ public abstract class Marshaller {
             boolean requireAllFields,
             boolean allowUnmappedFields) {
         if (mapper instanceof OneColumnMapper) {
-            return simpleMarshaller(cols, (OneColumnMapper<T>) mapper);
+            // Check a special case for key-only tables (i.e. views with Void value types)
+            if (cols.length == 0) {
+                if (mapper.targetType() == Void.class) {
+                    return new NoOpMarshaller();
+                } else {
+                    throw new IllegalArgumentException("Table is declared as key-only");
+                }
+            } else {
+                return simpleMarshaller(cols, (OneColumnMapper<T>) mapper);
+            }
         } else if (mapper instanceof PojoMapper) {
             return pojoMarshaller(cols, (PojoMapper<T>) mapper, requireAllFields, allowUnmappedFields);
         } else {
@@ -63,7 +72,7 @@ public abstract class Marshaller {
      * @param mapper Mapper.
      * @return Marshaller.
      */
-    static <T> SimpleMarshaller simpleMarshaller(MarshallerColumn[] cols, @NotNull OneColumnMapper<T> mapper) {
+    private static <T> SimpleMarshaller simpleMarshaller(MarshallerColumn[] cols, @NotNull OneColumnMapper<T> mapper) {
         final BinaryMode mode = MarshallerUtil.mode(mapper.targetType());
 
         final MarshallerColumn col = cols[0];
@@ -240,6 +249,22 @@ public abstract class Marshaller {
             for (int fldIdx = 0; fldIdx < fieldAccessors.length; fldIdx++) {
                 fieldAccessors[fldIdx].write(writer, obj);
             }
+        }
+    }
+
+    private static class NoOpMarshaller extends Marshaller {
+        @Override
+        public @Nullable Object value(Object obj, int fldIdx) {
+            return null;
+        }
+
+        @Override
+        public Object readObject(MarshallerReader reader, @Nullable Object target) {
+            return null;
+        }
+
+        @Override
+        public void writeObject(Object obj, MarshallerWriter writer) {
         }
     }
 }
